Time: 0 ms (100.00%) | Memory: 9.2 MB (21.87%) - LeetSync

+++ 8-string-to-integer-atoi/string-to-integer-atoi.cpp
@@ -4,7 +4,7 @@ public:
         int n = s.size();
 
         int sign = 1;
-        long long res = 0;
+        int res = 0;
 
         int i = 0;
         while (i < n && s[i] == ' ')
